pyproject.toml
[project]
name = "mimikopi"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "redtea" }
]
dependencies = [
    "pydub>=0.25.1",
    "matplotlib>=3.9.2",
    "scipy>=1.14.1",
    "japanize-matplotlib>=1.1.3",
    "opencv-python>=4.10.0.84",
    "moviepy>=1.0.3",
    "pyyaml>=6.0.2",
    "numpy>=2.1.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mimikopi"]
